Incorporate Ethereum-optimized Schnorr threshold signatures
Provides go code for generating threshold Schnorr signatunes, and solidity code
for on-chain verification of them.

The signatures are very cheap to verify on-chain (about 13k gas, if the signing
public key is already stored on-chain) due to a slight modification to the
Schnorr signature protocol which enables verification using ECRECOVER.

This implementation depends on "kyber" advanced-cryptography package from
EPFL-Dedis (which is unaffiliated with the Kyber Network, to the best of my
knowledge), and on btcsuite's "btcec" package for secp256k1 arithmetic.

The ethdss and ethschnorr packages and associated tests are modifications of the
kyber sign/dss and sign/schnorr packages, respectively.